Move ComicVine and GoogleBooks ExternalUrl providers to MediaBrowser.Providers.Books
