--- MediaBrowser.Providers/Plugins/ComicVine/Configuration/PluginConfiguration.cs
@@ -1,10 +1,0 @@
-using MediaBrowser.Model.Plugins;
-
-namespace MediaBrowser.Providers.Plugins.ComicVine;
-
-/// <summary>
-/// Plugin configuration for the Comic Vine provider.
-/// </summary>
-public class PluginConfiguration : BasePluginConfiguration
-{
-}
