--- MediaBrowser.Providers/Plugins/GoogleBooks/Configuration/PluginConfiguration.cs
@@ -1,10 +1,0 @@
-using MediaBrowser.Model.Plugins;
-
-namespace MediaBrowser.Providers.Plugins.GoogleBooks;
-
-/// <summary>
-/// Plugin configuration for the Google Books provider.
-/// </summary>
-public class PluginConfiguration : BasePluginConfiguration
-{
-}
